Move GridNetworkClient to grid.clients.network_client module
