+++ src/syft/grid/connections/http_client.py
@@ -2,29 +2,27 @@
 
 # third party
 import requests
-from nacl.signing import SigningKey
 
 # syft absolute
 import syft as sy
 from syft.core.common.message import SignedImmediateSyftMessageWithReply
 from syft.core.common.message import SignedImmediateSyftMessageWithoutReply
 from syft.core.common.message import SyftMessage
+from syft.core.common.serde.serializable import Serializable
 from syft.core.io.connection import ClientConnection
-from syft.core.io.route import SoloRoute
-from syft.core.node.network.client import NetworkClient
 
 
 class HTTPClient(ClientConnection):
-    def __init__(self):
+    def __init__(self) -> None:
         pass
-    
+
     def send_immediate_msg_with_reply(
         self, msg: SignedImmediateSyftMessageWithReply
-    ) -> requests.Response:
+    ) -> Serializable:
         blob = self.send_msg(msg=msg).text
-        if blob:
-            response = sy.deserialize(blob=blob, from_json=True)
-            return response
+
+        response = sy.deserialize(blob=blob, from_json=True)
+        return response
 
     def send_immediate_msg_without_reply(
         self, msg: SignedImmediateSyftMessageWithoutReply
@@ -44,25 +42,3 @@ class HTTPClient(ClientConnection):
     def _get_metadata(self, url: str) -> str:
         self.base_url = url
         return requests.get(url + "/metadata").text
-
-class GridNetworkClient(NetworkClient):
-
-    def __init__(self,url: str, conn: ClientConnection):
-        # generate a signing key
-        self.signing_key = SigningKey.generate()
-        self.verify_key = self.signing_key.verify_key
-
-        (
-            spec_location,
-            name,
-            client_id,
-        ) = NetworkClient.deserialize_client_metadata_from_node(metadata=conn._get_metadata(url))
-        route = SoloRoute(destination=spec_location, connection=conn)
-        print("\n\n\n\n name: ", name, "\n\n\n")
-        super().__init__(
-            network=spec_location,
-            name=name,
-            routes=[route],
-            signing_key=self.signing_key,
-            verify_key=self.verify_key,
-        )
